move GrpcXdsServer out of GrpcXdsBootstrap, and fix tests
