+++ build_autogenerated.yaml
@@ -1243,6 +1243,7 @@ libs:
   - src/core/xds/grpc/xds_route_config.h
   - src/core/xds/grpc/xds_route_config_parser.h
   - src/core/xds/grpc/xds_routing.h
+  - src/core/xds/grpc/xds_server_grpc.h
   - src/core/xds/grpc/xds_transport_grpc.h
   - src/core/xds/xds_client/xds_api.h
   - src/core/xds/xds_client/xds_bootstrap.h
@@ -2051,6 +2052,7 @@ libs:
   - src/core/xds/grpc/xds_route_config.cc
   - src/core/xds/grpc/xds_route_config_parser.cc
   - src/core/xds/grpc/xds_routing.cc
+  - src/core/xds/grpc/xds_server_grpc.cc
   - src/core/xds/grpc/xds_transport_grpc.cc
   - src/core/xds/xds_client/xds_api.cc
   - src/core/xds/xds_client/xds_bootstrap.cc
