+++ gRPC-C++.podspec
@@ -1352,6 +1352,7 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_route_config.h',
                       'src/core/xds/grpc/xds_route_config_parser.h',
                       'src/core/xds/grpc/xds_routing.h',
+                      'src/core/xds/grpc/xds_server_grpc.h',
                       'src/core/xds/grpc/xds_transport_grpc.h',
                       'src/core/xds/xds_client/xds_api.h',
                       'src/core/xds/xds_client/xds_bootstrap.h',
@@ -2633,6 +2634,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_route_config.h',
                               'src/core/xds/grpc/xds_route_config_parser.h',
                               'src/core/xds/grpc/xds_routing.h',
+                              'src/core/xds/grpc/xds_server_grpc.h',
                               'src/core/xds/grpc/xds_transport_grpc.h',
                               'src/core/xds/xds_client/xds_api.h',
                               'src/core/xds/xds_client/xds_bootstrap.h',
