+++ gRPC-Core.podspec
@@ -2121,6 +2121,8 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_route_config_parser.h',
                       'src/core/xds/grpc/xds_routing.cc',
                       'src/core/xds/grpc/xds_routing.h',
+                      'src/core/xds/grpc/xds_server_grpc.cc',
+                      'src/core/xds/grpc/xds_server_grpc.h',
                       'src/core/xds/grpc/xds_transport_grpc.cc',
                       'src/core/xds/grpc/xds_transport_grpc.h',
                       'src/core/xds/xds_client/xds_api.cc',
@@ -3410,6 +3412,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_route_config.h',
                               'src/core/xds/grpc/xds_route_config_parser.h',
                               'src/core/xds/grpc/xds_routing.h',
+                              'src/core/xds/grpc/xds_server_grpc.h',
                               'src/core/xds/grpc/xds_transport_grpc.h',
                               'src/core/xds/xds_client/xds_api.h',
                               'src/core/xds/xds_client/xds_bootstrap.h',
