+++ grpc.gemspec
@@ -2008,6 +2008,8 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/xds/grpc/xds_route_config_parser.h )
   s.files += %w( src/core/xds/grpc/xds_routing.cc )
   s.files += %w( src/core/xds/grpc/xds_routing.h )
+  s.files += %w( src/core/xds/grpc/xds_server_grpc.cc )
+  s.files += %w( src/core/xds/grpc/xds_server_grpc.h )
   s.files += %w( src/core/xds/grpc/xds_transport_grpc.cc )
   s.files += %w( src/core/xds/grpc/xds_transport_grpc.h )
   s.files += %w( src/core/xds/xds_client/xds_api.cc )
