+++ package.xml
@@ -1990,6 +1990,8 @@
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_route_config_parser.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_routing.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_routing.h" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_server_grpc.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_server_grpc.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_transport_grpc.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_transport_grpc.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/xds_client/xds_api.cc" role="src" />
