+++ src/core/xds/grpc/xds_transport_grpc.cc
@@ -254,7 +254,7 @@ class GrpcXdsTransportFactory::GrpcXdsTransport::StateWatcher final
 namespace {
 
 RefCountedPtr<Channel> CreateXdsChannel(
-    const ChannelArgs& args, const GrpcXdsBootstrap::GrpcXdsServer& server) {
+    const ChannelArgs& args, const GrpcXdsServer& server) {
   RefCountedPtr<grpc_channel_credentials> channel_creds =
       CoreConfiguration::Get().channel_creds_registry().CreateChannelCreds(
           server.channel_creds_config());
@@ -269,9 +269,8 @@ GrpcXdsTransportFactory::GrpcXdsTransport::GrpcXdsTransport(
     std::function<void(absl::Status)> on_connectivity_failure,
     absl::Status* status)
     : factory_(factory) {
-  channel_ = CreateXdsChannel(
-      factory->args_,
-      static_cast<const GrpcXdsBootstrap::GrpcXdsServer&>(server));
+  channel_ = CreateXdsChannel(factory->args_,
+                              static_cast<const GrpcXdsServer&>(server));
   CHECK(channel_ != nullptr);
   if (channel_->IsLame()) {
     *status = absl::UnavailableError("xds client has a lame channel");
