+++ test/core/xds/BUILD
@@ -36,6 +36,7 @@ grpc_cc_test(
     deps = [
         "//:gpr",
         "//src/core:grpc_xds_client",
+        "//src/core:xds_server_grpc",
         "//test/core/test_util:grpc_test_util",
         "//test/core/test_util:scoped_env_var",
     ],
