+++ test/core/xds/xds_lb_policy_registry_test.cc
@@ -76,8 +76,7 @@ absl::StatusOr<std::string> ConvertXdsPolicy(
   upb::Arena arena;
   upb::DefPool def_pool;
   XdsResourceType::DecodeContext context = {
-      nullptr, GrpcXdsBootstrap::GrpcXdsServer(), nullptr, def_pool.ptr(),
-      arena.ptr()};
+      nullptr, GrpcXdsServer(), nullptr, def_pool.ptr(), arena.ptr()};
   auto* upb_policy = envoy_config_cluster_v3_LoadBalancingPolicy_parse(
       serialized_policy.data(), serialized_policy.size(), arena.ptr());
   ValidationErrors errors;
